// FamilyClub.React/famiy.club.react/src/lib/api/generated/models/AuthorDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthorDTO
 */
export interface AuthorDTO {
    /**
     * 
     * @type {number}
     * @memberof AuthorDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof AuthorDTO
     */
    authorName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorDTO
     */
    biography?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorDTO
     */
    photoUrl?: string | null;
}

/**
 * Check if a given object implements the AuthorDTO interface.
 */
export function instanceOfAuthorDTO(value: object): value is AuthorDTO {
    return true;
}

export function AuthorDTOFromJSON(json: any): AuthorDTO {
    return AuthorDTOFromJSONTyped(json, false);
}

export function AuthorDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthorDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'authorName': json['authorName'] == null ? undefined : json['authorName'],
        'biography': json['biography'] == null ? undefined : json['biography'],
        'photoUrl': json['photoUrl'] == null ? undefined : json['photoUrl'],
    };
}

export function AuthorDTOToJSON(json: any): AuthorDTO {
    return AuthorDTOToJSONTyped(json, false);
}

export function AuthorDTOToJSONTyped(value?: AuthorDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'authorName': value['authorName'],
        'biography': value['biography'],
        'photoUrl': value['photoUrl'],
    };
}