// FamilyClub.React/famiy.club.react/src/lib/api/generated/models/ProductDto.ts
export interface ProductDto {
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ProductDto
     */
    productName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    discountPrice?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ProductDto
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    publisherId?: number | null;
    /**
     * 
     * @type {Array<ProductImage>}
     * @memberof ProductDto
     */
    productImages?: Array<ProductImage> | null;
    /**
     * 
     * @type {string}
     * @memberof ProductDto
     */
    originalTitle?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    pageCount?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof ProductDto
     */
    publishingDate?: Date | null;
    /**
     * 
     * @type {ProductFormat}
     * @memberof ProductDto
     */
    productFormat?: ProductFormat;
    /**
     * 
     * @type {CoverType}
     * @memberof ProductDto
     */
    coverType?: CoverType;
    /**
     * 
     * @type {Availability}
     * @memberof ProductDto
     */
    availability?: Availability;
    /**
     * 
     * @type {BookSize}
     * @memberof ProductDto
     */
    bookSize?: BookSize;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    quantityInStock?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    originalLanguageId?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ProductDto
     */
    isbn?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    promotionId?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ProductDto
     */
    productCode?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    weightGrams?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDto
     */
    itemsInSet?: number | null;
    /**
     * 
     * @type {AgeRestriction}
     * @memberof ProductDto
     */
    ageRestriction?: AgeRestriction;
    /**
     * 
     * @type {Array<number>}
     * @memberof ProductDto
     */
    authorIds?: Array<number> | null;
    /**
     * 
     * @type {Array<number>}
     * @memberof ProductDto
     */
    languageIds?: Array<number> | null;
    /**
     * 
     * @type {Array<number>}
     * @memberof ProductDto
     */
    categoryIds?: Array<number> | null;
    /**
     * 
     * @type {Array<number>}
     * @memberof ProductDto
     */
    seriesIds?: Array<number> | null;
    /**
     * 
     * @type {Array<number>}
     * @memberof ProductDto
     */
    translatorIds?: Array<number> | null;
    /**
     * 
     * @type {boolean}
     * @memberof ProductDto
     */
    leaveOldImages?: boolean;
}