// FamilyClub.React/famiy.club.react/src/lib/api/generated/models/ReviewDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ReviewDto
 */
export interface ReviewDto {
    /**
     * 
     * @type {number}
     * @memberof ReviewDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ReviewDto
     */
    productId?: number;
    /**
     * 
     * @type {string}
     * @memberof ReviewDto
     */
    userId?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ReviewDto
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof ReviewDto
     */
    comment?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ReviewDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof ReviewDto
     */
    approved?: boolean;
}

/**
 * Check if a given object implements the ReviewDto interface.
 */
export function instanceOfReviewDto(value: object): value is ReviewDto {
    return true;
}

export function ReviewDtoFromJSON(json: any): ReviewDto {
    return ReviewDtoFromJSONTyped(json, false);
}

export function ReviewDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReviewDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'productId': json['productId'] == null ? undefined : json['productId'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'rating': json['rating'] == null ? undefined : json['rating'],
        'comment': json['comment'] == null ? undefined : json['comment'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'approved': json['approved'] == null ? undefined : json['approved'],
    };
}

export function ReviewDtoToJSON(json: any): ReviewDto {
    return ReviewDtoToJSONTyped(json, false);
}

export function ReviewDtoToJSONTyped(value?: ReviewDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'productId': value['productId'],
        'userId': value['userId'],
        'rating': value['rating'],
        'comment': value['comment'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'approved': value['approved'],
    };
}