// FamilyClub.React/famiy.club.react/lib/api/generated/models/FormatDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FormatDto
 */
export interface FormatDto {
    /**
     * 
     * @type {number}
     * @memberof FormatDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof FormatDto
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FormatDto
     */
    code?: string | null;
}

/**
 * Check if a given object implements the FormatDto interface.
 */
export function instanceOfFormatDto(value: object): value is FormatDto {
    return true;
}

export function FormatDtoFromJSON(json: any): FormatDto {
    return FormatDtoFromJSONTyped(json, false);
}

export function FormatDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FormatDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'code': json['code'] == null ? undefined : json['code'],
    };
}

export function FormatDtoToJSON(json: any): FormatDto {
    return FormatDtoToJSONTyped(json, false);
}

export function FormatDtoToJSONTyped(value?: FormatDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'code': value['code'],
    };
}